+++ src/main/java/pl/smarthouse/module/ModuleCreatorType.java
@@ -1,0 +1,9 @@
+package pl.smarthouse.module;
+
+public enum ModuleCreatorType {
+  VENTILATION,
+  COMFORT,
+  WEATHER,
+  EXTERNAL_LIGHTS,
+  FIREPLACE
+}
